refactor(agent/story): 拆分 StoryAgentHost 为三段独立组件
将 1018 行的 story-agent.runtime.ts 中混在一起的 host facade、5 个 extension、
tool executor、工具函数全部拆开，主文件降至 358 行：

- StoryRuntimeTelemetry：所有 Dashboard 观测字段 + recordXxx，无业务依赖。
- StoryBatchPreparer：ledger → batch 数据流，独立维护游标和 finish 判定。
- StoryContextLifecycle：AgentContext snapshot 装载/回写 + 上下文压缩，
  仅向 telemetry 上报 retry 期间的可恢复错误。

StoryLoopAgent 取代旧的 StoryAgentHost facade，直接持有上述三个组件，
extension 也分别只依赖具体组件而非中间层。3 个单方法 telemetry extension
合并为 StoryTelemetryKernelExtension，剩余 2 个有完整语义的 extension
独立成文件，对齐 root-agent 已有的 extensions/ 布局。
